Split and import e_red_door data across all overlay
