+++ config/splat.us.stdre.yaml
@@ -40,7 +40,8 @@ segments:
       - [0x660, data]
       - [0x770, data]
       - [0x9A8, .data, collision]
-      - [0xD68, data]
+      - [0xD68, .data, e_red_door]
+      - [0xD80, data]
       - [0x1270, .data, entity_relic_orb]
       - [0x12D4, data]
       - [0x1498, roomdef, g_Rooms]
@@ -52,7 +53,7 @@ segments:
       - [0x1181C, rodata]
       - [0x118D0, rodata]
       - [0x11910, .rodata, 1697C] # EntityFadeToWhite
-      - [0x1193C, .rodata, 18B80] # EntityRedDoor
+      - [0x1193C, .rodata, e_red_door] # EntityRedDoor
       - [0x11954, rodata]
       - [0x119C4, .rodata, 1C7DC]
       - [0x119DC, .rodata, entity_relic_orb]
@@ -67,7 +68,9 @@ segments:
       - [0x16FC0, c, update]
       - [0x172BC, c, update_stage_entities]
       - [0x173C4, c, collision]
-      - [0x18B80, c]
+      - [0x18B80, c, create_entity]
+      - [0x196F8, c, e_red_door]
+      - [0x1A3A8, c]
       - [0x1C63C, c]
       - [0x1C7DC, c]
       - [0x1D074, c, entity_relic_orb]
