+++ config/splat.us.stno3.yaml
@@ -44,7 +44,8 @@ segments:
       - [0x11F0, data]
       - [0x1B40, data]
       - [0x1CEC, .data, collision]
-      - [0x20AC, data]
+      - [0x20AC, .data, e_red_door]
+      - [0x20C4, data]
       - [0x25B4, .data, entity_relic_orb]
       - [0x2618, data]
       - [0x3180, data]
@@ -66,7 +67,7 @@ segments:
       - [0x374F4, .rodata, 3FF00] # EntityDeathStolenItems
       - [0x37514, .rodata, 3FF00] # EntityDeath
       - [0x37564, .rodata, 3FF00] # EntityWargExplosionPuffOpaque
-      - [0x37578, .rodata, 4343C] # EntityRedDoor
+      - [0x37578, .rodata, e_red_door] # EntityRedDoor
       - [0x37590, rodata]
       - [0x375E0, .rodata, 46684] # EntityPrizeDrop
       - [0x37600, .rodata, 46684] # EntityEquipItemDrop
@@ -86,7 +87,9 @@ segments:
       - [0x4187C, c, update]
       - [0x41B78, c, update_stage_entities]
       - [0x41C80, c, collision]
-      - [0x4343C, c, 4343C]
+      - [0x4343C, c, create_entity]
+      - [0x43FB4, c, e_red_door]
+      - [0x44C64, c]
       - [0x46684, c]
       - [0x47930, c, entity_relic_orb]
       - [0x48400, c, entity_heart_drop]
