+++ src/st/cen/14264.c
@@ -1,12 +1,8 @@
-#include "st0.h"
-
-#include "../libstage.h"
-
-#include "../entity_red_door.h"
+#include "cen.h"
 
 #include "../entity.h"
 
-u8 func_801B4AF0(u8 frames[], Entity* self, u8 arg2) {
+u8 func_8019444C(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
     u8* var_s1 = &frames[animFrameStart];
     s16 var_a1 = 0;
@@ -38,7 +34,7 @@ u8 func_801B4AF0(u8 frames[], Entity* self, u8 arg2) {
 }
 
 // Absolute distance from g_CurrentEntity to the player in the X Axis
-s16 GetDistanceToPlayerX(void) {
+s32 GetDistanceToPlayerX(void) {
     s16 xDistance = g_CurrentEntity->posX.i.hi - PLAYER.posX.i.hi;
 
     if (xDistance < 0) {
@@ -57,16 +53,9 @@ s32 GetDistanceToPlayerY(void) {
     return yDistance;
 }
 
-s32 GetSideToPlayer(void) {
-    s16 side = g_CurrentEntity->posX.i.hi > PLAYER.posX.i.hi;
+#include "../get_side_to_player.h"
 
-    if (g_CurrentEntity->posY.i.hi > PLAYER.posY.i.hi) {
-        side |= 2;
-    }
-    return side;
-}
-
-void MoveEntity(void) {
+void MoveEntity() {
     g_CurrentEntity->posX.val += g_CurrentEntity->velocityX;
     g_CurrentEntity->posY.val += g_CurrentEntity->velocityY;
 }
@@ -77,23 +66,7 @@ void FallEntity(void) {
     }
 }
 
-u8 func_801B4D18(void) {
-    u8 unkState;
-    Entity* entity;
-
-    MoveEntity();
-    FallEntity();
-
-    entity = g_CurrentEntity;
-
-    if (unkState & 1) {
-        entity->velocityY = 0;
-    }
-
-    return unkState;
-}
-
-s32 func_801B4D5C(u16* sensors) {
+s32 func_80194674(u16* sensors) {
     Collider col;
     Collider colBack;
     s16 x;
@@ -146,53 +119,107 @@ s32 func_801B4D5C(u16* sensors) {
     return 0;
 }
 
-INCLUDE_ASM("st/st0/nonmatchings/330E0", func_801B4FD4);
+s32 func_801948EC(s16* posX) {
+    Collider collider;
+    s16 temp2;
+    s16 temp4;
+    s16 x, y;
 
-INCLUDE_ASM("st/st0/nonmatchings/330E0", func_801B51E4);
+    g_CurrentEntity->posX.val += g_CurrentEntity->velocityX;
+    temp2 = g_CurrentEntity->posY.i.hi + 3;
+    g_CurrentEntity->posY.i.hi = temp2;
+    x = g_CurrentEntity->posX.i.hi + *posX;
+    posX++;
+    y = temp2 + *posX;
+    g_api.CheckCollision(x, y, &collider, 0);
+    if (!(collider.effects & EFFECT_SOLID)) {
+        return 0;
+    }
+    posX++;
+
+    g_CurrentEntity->posY.i.hi = g_CurrentEntity->posY.i.hi + collider.unk18;
+    if (g_CurrentEntity->velocityX != 0) {
+        if (g_CurrentEntity->velocityX < 0) {
+            temp4 = x - *posX;
+            posX++;
+        } else {
+            temp4 = x + *posX;
+            posX++;
+        }
+        y = y + *posX - 7;
+        g_api.CheckCollision(temp4, y, &collider, 0);
+        if (collider.effects & EFFECT_SOLID) {
+            if ((collider.effects & (EFFECT_UNK_8000 | EFFECT_UNK_0002)) ==
+                EFFECT_UNK_0002) {
+                g_CurrentEntity->posX.val =
+                    g_CurrentEntity->posX.val - g_CurrentEntity->velocityX;
+                g_CurrentEntity->velocityX = 0;
+                return 0xFF;
+            }
+            return 0x61;
+        }
+        y += 15;
+        g_api.CheckCollision(temp4, y, &collider, 0);
+        if (collider.effects & EFFECT_SOLID) {
+            if (collider.effects & EFFECT_UNK_8000) {
+                return 0x61;
+            }
+            return 1;
+        }
+        g_CurrentEntity->posX.val -= g_CurrentEntity->velocityX;
+        g_CurrentEntity->velocityX = 0;
+
+        return 0x80;
+    }
+    return 1;
+}
 
 Entity* AllocEntity(Entity* start, Entity* end) {
     Entity* current = start;
+
     while (current < end) {
         if (current->entityId == E_NONE) {
             DestroyEntity(current);
             return current;
         }
-
         current++;
     }
     return NULL;
 }
 
-s32 func_801B542C(u8 arg0, s16 arg1) { return D_80181990[arg0] * arg1; }
+s32 func_80194B34(u8 arg0, s16 arg1) { return D_80180BBC[arg0] * arg1; }
 
-s16 func_801B5458(u8 arg0) { return D_80181990[arg0]; }
+s16 func_80194B60(u8 arg0) { return D_80180BBC[arg0]; }
 
-void func_801B5474(s32 arg0, s16 arg1) {
-    g_CurrentEntity->velocityX = func_801B542C(arg0, arg1);
-    g_CurrentEntity->velocityY = func_801B542C(arg0 - 0x40, arg1);
+void func_80194B7C(s32 arg0, s16 arg1) {
+    g_CurrentEntity->velocityX = func_80194B34(arg0 & 0xFF, arg1);
+    g_CurrentEntity->velocityY = func_80194B34((arg0 - 0x40) & 0xFF, arg1);
 }
 
-u8 func_801B54E0(s16 arg0, s16 arg1) {
-    return (ratan2(arg1, arg0) >> 4) + 0x40;
+u8 func_80194BE8(s16 x, s16 y) { return (ratan2(y, x) >> 4) + 0x40; }
+
+u8 func_80194C20(Entity* arg0, Entity* arg1) {
+    u16 x;
+    u16 y;
+
+    x = arg1->posX.i.hi - arg0->posX.i.hi;
+    y = arg1->posY.i.hi - arg0->posY.i.hi;
+
+    return func_80194BE8(x, y);
 }
 
-u8 func_801B5518(Entity* arg0, Entity* arg1) {
-    s16 a = arg1->posX.i.hi - arg0->posX.i.hi;
-    s16 b = arg1->posY.i.hi - arg0->posY.i.hi;
-    return func_801B54E0(a, b);
-}
+u16 func_80194C68(s16 x, s16 y) {
+    x -= g_CurrentEntity->posX.i.hi;
+    y -= g_CurrentEntity->posY.i.hi;
 
-u8 func_801B5560(s32 arg0, s32 arg1) {
-    s16 a = (arg0 - (u16)g_CurrentEntity->posX.i.hi);
-    s16 b = (arg1 - (u16)g_CurrentEntity->posY.i.hi);
-    return func_801B54E0(a, b);
+    return func_80194BE8(x, y);
 }
 
 #include "../adjust_value_within_threshold.h"
 
 #include "../unk_entity_func0.h"
 
-u16 func_801B568C(s16 x, s16 y) { return ratan2(y, x); }
+u16 func_80194D94(s16 arg0, s16 arg1) { return ratan2(arg1, arg0); }
 
 u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     s32 diffX = b->posX.i.hi - a->posX.i.hi;
@@ -200,74 +227,82 @@ u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     return ratan2(diffY, diffX);
 }
 
-u16 func_801B56F4(s32 x, s32 y) {
+u16 func_80194DFC(s32 x, s32 y) {
     s16 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
     s16 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
     return ratan2(diffY, diffX);
 }
 
-u16 GetNormalizedAngle(u16 arg0, s16 arg1, s16 arg2) {
+u16 GetNormalizedAngle(u16 arg0, u16 arg1, u16 arg2) {
+    u16 var_v0 = arg1;
     u16 temp_a2 = arg2 - arg1;
-    u16 ret;
+    u16 var_v0_2;
 
     if (temp_a2 & 0x800) {
-        ret = temp_a2 & 0x7FF;
+        var_v0_2 = (0x800 - temp_a2) & 0x7FF;
     } else {
-        ret = temp_a2;
+        var_v0_2 = temp_a2;
     }
 
-    if (ret > arg0) {
+    if (var_v0_2 > arg0) {
         if (temp_a2 & 0x800) {
-            ret = arg1 - arg0;
+            var_v0 = arg1 - arg0;
         } else {
-            ret = arg1 + arg0;
+            var_v0 = arg1 + arg0;
         }
 
-        return ret;
+        return var_v0;
     }
 
     return arg2;
 }
 
 void SetStep(u8 step) {
-    g_CurrentEntity->step = step;
-    g_CurrentEntity->step_s = 0;
-    g_CurrentEntity->animFrameIdx = 0;
-    g_CurrentEntity->animFrameDuration = 0;
+    Entity* entity = g_CurrentEntity;
+
+    entity->step = step;
+    entity->step_s = 0;
+    entity->animFrameIdx = 0;
+    entity->animFrameDuration = 0;
 }
 
-void SetSubStep(u8 step_s) {
-    g_CurrentEntity->step_s = step_s;
-    g_CurrentEntity->animFrameIdx = 0;
-    g_CurrentEntity->animFrameDuration = 0;
+void func_80194EC4(u8 arg0) {
+    Entity* entity = g_CurrentEntity;
+
+    entity->step_s = arg0;
+    entity->animFrameIdx = 0;
+    entity->animFrameDuration = 0;
 }
 
-void func_801B57D0(u16 params) {
-    Entity* current;
+void EntityExplosionSpawn(u16 arg0, u16 arg1) {
+    Entity* entity;
 
-    if (params == 0xFF) {
+    if (arg1 != 0) {
+        func_8019A328(arg1);
+    }
+    if (arg0 == 0xFF) {
         DestroyEntity(g_CurrentEntity);
         return;
     }
-    current = g_CurrentEntity;
-    g_CurrentEntity->entityId = E_EXPLOSION;
-    g_CurrentEntity->pfnUpdate = EntityExplosion;
-    g_CurrentEntity->drawFlags = 0;
-    current->params = params;
-    current->animCurFrame = 0;
+    entity = g_CurrentEntity;
+    entity->drawFlags = 0;
+    entity->entityId = 2;
+    entity->pfnUpdate = EntityExplosion;
+    entity->params = arg0;
+    entity->animCurFrame = 0;
     g_CurrentEntity->step = 0;
     g_CurrentEntity->step_s = 0;
 }
 
 #include "../init_entity.h"
 
-void EntityDummy(Entity* arg0) {
-    if (arg0->step == 0) {
-        arg0->step++;
+void func_80195070(Entity* entity) {
+    if (entity->step == 0) {
+        entity->step++;
     }
 }
 
-s32 func_801B5970(u16* hitSensors, s16 sensorCount) {
+s32 func_80195098(u16* hitSensors, s16 sensorCount) {
     Collider collider;
     s16 i;
     s32 velocityX;
@@ -290,8 +325,8 @@ s32 func_801B5970(u16* hitSensors, s16 sensorCount) {
 
             y += *hitSensors++;
             g_api.CheckCollision(x, y, &collider, 0);
-            if (collider.effects & 2 &&
-                ((!(collider.effects & 0x8000)) || (i != 0))) {
+            if (collider.effects & EFFECT_UNK_0002 &&
+                ((!(collider.effects & EFFECT_UNK_8000)) || (i != 0))) {
                 return 2;
             }
         }
@@ -299,7 +334,7 @@ s32 func_801B5970(u16* hitSensors, s16 sensorCount) {
     }
 }
 
-void func_801B5A98(u16* hitSensors, s16 sensorCount) {
+void func_801951C0(u16* hitSensors, s16 sensorCount) {
     Collider collider;
     s16 i;
     s32 velocityX;
@@ -320,7 +355,8 @@ void func_801B5A98(u16* hitSensors, s16 sensorCount) {
 
         y += *hitSensors++;
         g_api.CheckCollision(x, y, &collider, 0);
-        if (collider.effects & 2 && (!(collider.effects & 0x8000) || i != 0)) {
+        if (collider.effects & EFFECT_UNK_0002 &&
+            (!(collider.effects & EFFECT_UNK_8000) || i != 0)) {
             if (velocityX < 0) {
                 g_CurrentEntity->posX.i.hi += collider.unk1C;
             } else {
@@ -331,11 +367,11 @@ void func_801B5A98(u16* hitSensors, s16 sensorCount) {
     }
 }
 
-INCLUDE_ASM("st/st0/nonmatchings/330E0", func_801B5BF0);
+#include "../get_player_collision_with.h"
 
 #include "../replace_breakable_with_item_drop.h"
 
-void func_801B5EC8(void) {
+void func_80195714(void) {
     s32 temp_v1;
     Entity* entity;
 
@@ -357,22 +393,22 @@ void func_801B5EC8(void) {
     }
 }
 
-void func_801B5F4C(u16 arg0) {
-    Collider res;
+void func_80195798(u16 arg0) {
+    Collider collider;
 
     if (g_CurrentEntity->velocityX < 0) {
         g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
-                             g_CurrentEntity->posY.i.hi - 7, &res, 0);
-        if (res.effects & 5) {
+                             g_CurrentEntity->posY.i.hi - 7, &collider, 0);
+        if (collider.effects & EFFECT_NOTHROUGH) {
             g_CurrentEntity->velocityY = 0;
         }
     }
 
-    g_api.CheckCollision(
-        g_CurrentEntity->posX.i.hi, g_CurrentEntity->posY.i.hi + 7, &res, 0);
+    g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
+                         g_CurrentEntity->posY.i.hi + 7, &collider, 0);
 
     if (arg0) {
-        if (!(res.effects & 5)) {
+        if (!(collider.effects & EFFECT_NOTHROUGH)) {
             MoveEntity();
             FallEntity();
             return;
@@ -381,40 +417,24 @@ void func_801B5F4C(u16 arg0) {
         g_CurrentEntity->velocityX = 0;
         g_CurrentEntity->velocityY = 0;
 
-        if (res.effects & 4) {
+        if (collider.effects & EFFECT_QUICKSAND) {
             g_CurrentEntity->posY.val += FIX(0.125);
             return;
         }
 
-        g_CurrentEntity->posY.i.hi += res.unk18;
+        g_CurrentEntity->posY.i.hi += collider.unk18;
         return;
     }
 
-    if (!(res.effects & 5)) {
+    if (!(collider.effects & EFFECT_NOTHROUGH)) {
         MoveEntity();
-        func_801B5EC8();
+        func_80195714();
     }
 }
 
-// Can't be deduped with the other instances due to the extra memcpy
-void CollectHeart(u16 index) {
-    s8 hearts[10];
-
-    __builtin_memcpy(&hearts[0], &c_HeartPrizes[0], 2);
-    __builtin_memcpy(&hearts[8], &c_HeartPrizes[4], 2);
-    g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-    g_Status.hearts = hearts[index] + g_Status.hearts;
-    if (g_Status.heartsMax < g_Status.hearts) {
-        g_Status.hearts = g_Status.heartsMax;
-    }
-    DestroyEntity(g_CurrentEntity);
-}
+#include "../collect_heart.h"
 
 #include "../collect_gold.h"
-
-void func_801B61D4(void) {}
-
-void func_801B61DC(void) { DestroyEntity(g_CurrentEntity); }
 
 void CollectSubweapon(u16 subWeaponIdx) {
     Entity* player = &PLAYER;
@@ -422,21 +442,20 @@ void CollectSubweapon(u16 subWeaponIdx) {
 
     g_api.PlaySfx(NA_SE_PL_IT_PICKUP);
     subWeapon = g_Status.subWeapon;
-    g_Status.subWeapon = D_80181CAC[subWeaponIdx];
+    g_Status.subWeapon = D_80180EEC[subWeaponIdx];
 
     if (subWeapon == g_Status.subWeapon) {
         subWeapon = 1;
         g_CurrentEntity->unk6D[0] = 0x10;
     } else {
-        subWeapon = D_80181CDC[subWeapon];
+        subWeapon = D_80180F1C[subWeapon];
         g_CurrentEntity->unk6D[0] = 0x60;
     }
 
     if (subWeapon != 0) {
         g_CurrentEntity->params = subWeapon;
         g_CurrentEntity->posY.i.hi = player->posY.i.hi + 12;
-        g_CurrentEntity->step = 7;
-        g_CurrentEntity->step_s = 0;
+        SetStep(7);
         g_CurrentEntity->velocityY = FIX(-2.5);
         g_CurrentEntity->animCurFrame = 0;
         g_CurrentEntity->ext.generic.unk88.S16.unk2 = 5;
@@ -450,10 +469,24 @@ void CollectSubweapon(u16 subWeaponIdx) {
     DestroyEntity(g_CurrentEntity);
 }
 
-void CollectDummy(void) { DestroyEntity(g_CurrentEntity); }
+#include "../collect_heart_vessel.h"
 
-Entity* func_801B633C(void) {
-    g_CurrentEntity->step = 3;
-    g_CurrentEntity->params = 4;
-    return g_CurrentEntity;
+void func_80195C0C(void) {
+    g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
+    g_api.func_800FE044(5, 0x8000);
+    DestroyEntity(g_CurrentEntity);
 }
+
+void func_80195C5C(void) { DestroyEntity(g_CurrentEntity); }
+
+INCLUDE_ASM("st/cen/nonmatchings/14264", EntityPrizeDrop);
+
+#define MISSING_ANIMATE_ENTITY_PROTOTYPE
+#include "../entity_explosion.h"
+#undef MISSING_ANIMATE_ENTITY_PROTOTYPE
+
+#include "../blink_item.h"
+
+INCLUDE_ASM("st/cen/nonmatchings/14264", EntityEquipItemDrop);
+
+#include "../blit_char.h"
