+++ src/st/cen/create_entity.c
@@ -1,0 +1,1 @@
+#include "../libstage.h"
