+++ src/st/cen/e_red_door.c
@@ -1,3 +1,1 @@
-#include "wrp.h"
-
 #include "../e_red_door.h"
