+++ src/st/dre/1A3A8.c
@@ -1,50 +1,46 @@
-#include "np3.h"
-
-#include "../libstage.h"
-
-#include "../entity_red_door.h"
+#include "dre.h"
 
 #include "../entity.h"
 
-u8 func_801BC6BC(u8 frames[], Entity* self, u8 arg2) {
+u8 func_8019A590(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
-    u8* iterEnt1 = &frames[animFrameStart];
+    u8* var_s1 = &frames[animFrameStart];
     s16 var_a1 = 0;
 
     if (self->animFrameDuration == 0) {
-        if (*iterEnt1 != 0) {
-            if (*iterEnt1 == 0xFF) {
+        if (*var_s1 != 0) {
+            if (*var_s1 == 0xFF) {
                 return 0;
             }
-            self->animFrameDuration = *iterEnt1++ + (u8)self->ext.stub[0x3F];
-            self->animCurFrame = *iterEnt1++;
+            self->animFrameDuration = *var_s1++ + (u8)self->ext.stub[0x3F];
+            self->animCurFrame = *var_s1++;
             self->animFrameIdx++;
             var_a1 = 128;
         } else {
-            iterEnt1 = frames;
+            var_s1 = frames;
             self->animFrameIdx = 0;
             self->animFrameDuration = 0;
             self->ext.stub[0x3F] = (arg2 * Random()) >> 8;
-            self->animFrameDuration = *iterEnt1++ + (u8)self->ext.stub[0x3F];
-            self->animCurFrame = *iterEnt1;
+            self->animFrameDuration = *var_s1++ + (u8)self->ext.stub[0x3F];
+            self->animCurFrame = *var_s1;
             self->animFrameIdx++;
             return 0;
         }
     }
     self->animFrameDuration--;
-    self->animCurFrame = iterEnt1[-1];
+    self->animCurFrame = var_s1[-1];
     var_a1 |= 1;
     return var_a1;
 }
 
 // Absolute distance from g_CurrentEntity to the player in the X Axis
 s32 GetDistanceToPlayerX(void) {
-    s16 yDistance = g_CurrentEntity->posX.i.hi - PLAYER.posX.i.hi;
+    s16 xDistance = g_CurrentEntity->posX.i.hi - PLAYER.posX.i.hi;
 
-    if (yDistance < 0) {
-        yDistance = -yDistance;
+    if (xDistance < 0) {
+        xDistance = -xDistance;
     }
-    return yDistance;
+    return xDistance;
 }
 
 // Absolute distance from g_CurrentEntity to the player in the Y Axis
@@ -57,7 +53,14 @@ s32 GetDistanceToPlayerY(void) {
     return yDistance;
 }
 
-#include "../get_side_to_player.h"
+s32 GetSideToPlayer(void) {
+    s16 var_a0 = g_CurrentEntity->posX.i.hi > PLAYER.posX.i.hi;
+
+    if (g_CurrentEntity->posY.i.hi > PLAYER.posY.i.hi) {
+        var_a0 |= 2;
+    }
+    return var_a0;
+}
 
 void MoveEntity(void) {
     g_CurrentEntity->posX.val += g_CurrentEntity->velocityX;
@@ -70,7 +73,7 @@ void FallEntity(void) {
     }
 }
 
-s32 func_801BC8E4(u16* sensors) {
+s32 func_8019A7B8(u16* sensors) {
     Collider col;
     Collider colBack;
     s16 x;
@@ -123,7 +126,7 @@ s32 func_801BC8E4(u16* sensors) {
     return 0;
 }
 
-s32 func_801BCB5C(s16* posX) {
+s32 func_8019AA30(s16* posX) {
     Collider collider;
     s16 temp2;
     s16 temp4;
@@ -186,39 +189,47 @@ Entity* AllocEntity(Entity* start, Entity* end) {
             DestroyEntity(current);
             return current;
         }
+
         current++;
     }
     return NULL;
 }
 
-s32 func_801BCDA4(u8 arg0, s16 arg1) { return D_80181A50[arg0] * arg1; }
+s32 func_8019AC78(u8 arg0, s16 arg1) { return D_80180D80[arg0] * arg1; }
 
-s16 func_801BCDD0(u8 arg0) { return D_80181A50[arg0]; }
+extern s16 D_80180D80[];
+s16 func_8019ACA4(u8 arg0) { return D_80180D80[arg0]; }
 
-void func_801BCDEC(s32 arg0, s16 arg1) {
-    g_CurrentEntity->velocityX = func_801BCDA4(arg0, arg1);
-    g_CurrentEntity->velocityY = func_801BCDA4(arg0 - 0x40, arg1);
+void func_8019ACC0(s32 arg0, s16 arg1) {
+    g_CurrentEntity->velocityX = func_8019AC78(arg0, arg1);
+    g_CurrentEntity->velocityY = func_8019AC78(arg0 - 0x40, arg1);
 }
 
-u8 func_801BCE58(s16 x, s16 y) { return (ratan2(y, x) >> 4) + 0x40; }
+u8 func_8019AD2C(s16 x, s16 y) { return (ratan2(y, x) >> 4) + 0x40; }
 
-u8 func_801BCE90(Entity* a, Entity* b) {
-    s32 diffX = (u16)b->posX.i.hi - (u16)a->posX.i.hi;
-    s32 diffY = (u16)b->posY.i.hi - (u16)a->posY.i.hi;
-    return func_801BCE58(diffX, diffY);
+u8 func_8019AD64(ObjInit2* arg0, ObjInit2* arg1) {
+    u16 x, y;
+
+    x = arg1->zPriority - arg0->zPriority;
+    y = arg1->palette - arg0->palette;
+
+    return func_8019AD2C(x, y);
 }
 
-u8 func_801BCED8(s32 x, s32 y) {
-    s32 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
-    s32 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
-    return func_801BCE58(diffX, diffY);
+u8 func_8019ADAC(s16 arg0, s16 arg1) {
+    s16 x, y;
+
+    x = arg0 - g_CurrentEntity->posX.i.hi;
+    y = arg1 - g_CurrentEntity->posY.i.hi;
+
+    return func_8019AD2C(x, y);
 }
 
 #include "../adjust_value_within_threshold.h"
 
 #include "../unk_entity_func0.h"
 
-u16 func_801BD004(s16 x, s16 y) { return ratan2(y, x); }
+u16 func_8019AED8(s16 arg0, s16 arg1) { return ratan2(arg1, arg0); }
 
 u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     s32 diffX = b->posX.i.hi - a->posX.i.hi;
@@ -226,7 +237,7 @@ u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     return ratan2(diffY, diffX);
 }
 
-u16 func_801BD06C(s32 x, s32 y) {
+u16 func_8019AF40(s32 x, s32 y) {
     s16 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
     s16 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
     return ratan2(diffY, diffX);
@@ -257,21 +268,29 @@ u16 GetNormalizedAngle(u16 arg0, s16 arg1, s16 arg2) {
 }
 
 void SetStep(u8 step) {
-    g_CurrentEntity->step = step;
-    g_CurrentEntity->step_s = 0;
-    g_CurrentEntity->animFrameIdx = 0;
-    g_CurrentEntity->animFrameDuration = 0;
+    Entity* entity;
+
+    entity = g_CurrentEntity;
+    entity->step = step;
+    entity->step_s = 0;
+    entity->animFrameIdx = 0;
+    entity->animFrameDuration = 0;
 }
 
 void SetSubStep(u8 step_s) {
-    g_CurrentEntity->step_s = step_s;
-    g_CurrentEntity->animFrameIdx = 0;
-    g_CurrentEntity->animFrameDuration = 0;
+    Entity* entity;
+
+    entity = g_CurrentEntity;
+    entity->step_s = step_s;
+    entity->animFrameIdx = 0;
+    entity->animFrameDuration = 0;
 }
 
-void EntityExplosionSpawn(u16 arg0, u16 sfxId) {
-    if (sfxId != 0) {
-        func_801C2598(sfxId);
+void EntityExplosionSpawn(u16 arg0, u16 arg1) {
+    Entity* entity;
+
+    if (arg1 != 0) {
+        func_801A046C(arg1);
     }
 
     if (arg0 == 0xFF) {
@@ -279,31 +298,33 @@ void EntityExplosionSpawn(u16 arg0, u16 sfxId) {
         return;
     }
 
-    g_CurrentEntity->entityId = E_EXPLOSION;
-    g_CurrentEntity->pfnUpdate = (PfnEntityUpdate)EntityExplosion;
-    g_CurrentEntity->params = arg0;
-    g_CurrentEntity->animCurFrame = 0;
-    g_CurrentEntity->drawFlags = 0;
+    entity = g_CurrentEntity;
+    entity->drawFlags = 0;
+    entity->entityId = E_EXPLOSION;
+    entity->pfnUpdate = (PfnEntityUpdate)EntityExplosion;
+    entity->params = arg0;
+    entity->animCurFrame = 0;
     g_CurrentEntity->step = 0;
     g_CurrentEntity->step_s = 0;
 }
 
 #include "../init_entity.h"
 
-void EntityDummy(Entity* arg0) {
+void func_8019B1B4(Entity* arg0) {
     if (arg0->step == 0) {
         arg0->step++;
     }
 }
 
-s32 func_801BD308(u16* hitSensors, s16 sensorCount) {
-    s32 velocityX = g_CurrentEntity->velocityX;
+s32 func_8019B1DC(u16* hitSensors, s16 sensorCount) {
     Collider collider;
+    s16 i;
+    s32 velocityX;
     u16 temp_a1;
     s16 x;
     s16 y;
-    s16 i;
 
+    velocityX = g_CurrentEntity->velocityX;
     if (velocityX != 0) {
         x = g_CurrentEntity->posX.i.hi;
         y = g_CurrentEntity->posY.i.hi;
@@ -327,19 +348,18 @@ s32 func_801BD308(u16* hitSensors, s16 sensorCount) {
     }
 }
 
-void func_801BD430(u16* hitSensors, s16 sensorCount) {
-    s32 velocityX = g_CurrentEntity->velocityX;
+void func_8019B304(u16* hitSensors, s16 sensorCount) {
     Collider collider;
+    s16 i;
+    s32 velocityX;
     s16 x;
     s16 y;
-    s16 i;
 
+    velocityX = g_CurrentEntity->velocityX;
     if (velocityX == 0)
         return;
-
     x = g_CurrentEntity->posX.i.hi;
     y = g_CurrentEntity->posY.i.hi;
-
     for (i = 0; i < sensorCount; i++) {
         if (velocityX < 0) {
             x = x + *hitSensors++;
@@ -365,19 +385,19 @@ void func_801BD430(u16* hitSensors, s16 sensorCount) {
 
 #include "../replace_breakable_with_item_drop.h"
 
-void func_801BD984(void) {
+void func_8019B858(void) {
     s32 temp_v1;
-    Entity* entity;
+    Entity* entity = g_CurrentEntity;
 
-    entity = g_CurrentEntity;
     if (entity->velocityY >= 0) {
         temp_v1 =
-            entity->ext.generic.unk88.S16.unk0 + entity->ext.generic.unk84.unk;
+            entity->ext.generic.unk84.unk + entity->ext.generic.unk88.S16.unk0;
         entity->ext.generic.unk84.unk = temp_v1;
         entity->velocityX = temp_v1;
-        if (temp_v1 == 0x10000 || temp_v1 == -0x10000) {
-            entity->ext.generic.unk88.S16.unk0 =
-                -entity->ext.generic.unk88.S16.unk0;
+
+        if ((temp_v1 == 0x10000) || (temp_v1 == -0x10000)) {
+            entity->ext.generic.unk88.U16.unk0 =
+                -entity->ext.generic.unk88.U16.unk0;
         }
         entity = g_CurrentEntity;
     }
@@ -387,7 +407,7 @@ void func_801BD984(void) {
     }
 }
 
-void func_801BDA08(u16 arg0) {
+void func_8019B8DC(u16 arg0) {
     Collider collider;
 
     if (g_CurrentEntity->velocityX < 0) {
@@ -413,18 +433,32 @@ void func_801BDA08(u16 arg0) {
 
         if (collider.effects & EFFECT_QUICKSAND) {
             g_CurrentEntity->posY.val += FIX(0.125);
-        } else {
-            g_CurrentEntity->posY.i.hi += collider.unk18;
+            return;
         }
-    } else {
-        if (!(collider.effects & EFFECT_NOTHROUGH)) {
-            MoveEntity();
-            func_801BD984();
-        }
+
+        g_CurrentEntity->posY.i.hi += collider.unk18;
+        return;
+    }
+
+    if (!(collider.effects & EFFECT_NOTHROUGH)) {
+        MoveEntity();
+        func_8019B858();
     }
 }
 
-#include "../collect_heart.h"
+void func_8019BA38(u16 arg0) {
+    s32* hearts;
+
+    g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
+    hearts = &g_Status.hearts;
+    *hearts += c_HeartPrizes[arg0];
+
+    if (g_Status.heartsMax < *hearts) {
+        *hearts = g_Status.heartsMax;
+    }
+
+    DestroyEntity(g_CurrentEntity);
+}
 
 #include "../collect_gold.h"
 
@@ -434,13 +468,13 @@ void CollectSubweapon(u16 subWeaponIdx) {
 
     g_api.PlaySfx(NA_SE_PL_IT_PICKUP);
     subWeapon = g_Status.subWeapon;
-    g_Status.subWeapon = D_80181D80[subWeaponIdx];
+    g_Status.subWeapon = D_801810B0[subWeaponIdx];
 
     if (subWeapon == g_Status.subWeapon) {
         subWeapon = 1;
         g_CurrentEntity->unk6D[0] = 0x10;
     } else {
-        subWeapon = D_80181DB0[subWeapon];
+        subWeapon = D_801810E0[subWeapon];
         g_CurrentEntity->unk6D[0] = 0x60;
     }
 
@@ -465,8 +499,10 @@ void CollectSubweapon(u16 subWeaponIdx) {
 
 void CollectLifeVessel(void) {
     g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-    g_api.func_800FE044(5, 0x8000);
+    g_api.func_800FE044(LIFE_VESSEL_INCREASE, 0x8000);
     DestroyEntity(g_CurrentEntity);
 }
 
 void DestroyCurrentEntity(void) { DestroyEntity(g_CurrentEntity); }
+
+INCLUDE_ASM("st/dre/nonmatchings/1A3A8", EntityPrizeDrop);
