+++ src/st/e_red_door.h
@@ -1,9 +1,35 @@
-#include <game.h>
+#include <stage.h>
 
-#include "entity_is_near_player.h"
-
-extern u8 g_eRedDoorUV[3][8];
+extern u16 g_eInitGeneric2[];
 extern u16 g_eRedDoorTiles[2][8];
+
+u8 g_eRedDoorUV[][8] = {
+    {0xB1, 0xB7, 0xB1, 0xB7, 0x21, 0x21, 0x5F, 0x5F},
+    {0x88, 0xA8, 0x88, 0xA8, 0x21, 0x21, 0x5F, 0x5F},
+    {0xA8, 0x88, 0xA8, 0x88, 0x21, 0x21, 0x5F, 0x5F},
+};
+
+static bool EntityIsNearPlayer(Entity* self) {
+    s16 distanceX;
+    s16 diffX;
+    s16 distanceY;
+    s16 diffY;
+
+    diffX = PLAYER.posX.i.hi - self->posX.i.hi;
+    distanceX = abs(diffX);
+    if (distanceX > 16) {
+        return false;
+    }
+
+    diffY = PLAYER.posY.i.hi - self->posY.i.hi;
+    distanceY = abs(diffY);
+    if (distanceY > 32) {
+        return false;
+    }
+
+    return true;
+}
+
 void EntityRedDoor(Entity* self) {
     Primitive* prim;
     s32 i;
