--- src/st/entity_is_near_player.h
@@ -1,20 +1,0 @@
-bool EntityIsNearPlayer(Entity* self) {
-    s16 distanceX;
-    s16 diffX;
-    s16 distanceY;
-    s16 diffY;
-
-    diffX = PLAYER.posX.i.hi - self->posX.i.hi;
-    distanceX = abs(diffX);
-    if (distanceX > 16) {
-        return false;
-    }
-
-    diffY = PLAYER.posY.i.hi - self->posY.i.hi;
-    distanceY = abs(diffY);
-    if (distanceY > 32) {
-        return false;
-    }
-
-    return true;
-}
