+++ src/st/mad/e_red_door.c
@@ -1,3 +1,3 @@
-#include "wrp.h"
+#include "mad.h"
 
 #include "../e_red_door.h"
