+++ src/st/no3/create_entity.c
@@ -1,0 +1,1 @@
+#include "../libstage.h"
