+++ src/st/np3/3C4D4.c
@@ -1,50 +1,46 @@
-#include "cen.h"
-
-#include "../libstage.h"
-
-#include "../entity_red_door.h"
+#include "np3.h"
 
 #include "../entity.h"
 
-u8 func_8019444C(u8 frames[], Entity* self, u8 arg2) {
+u8 func_801BC6BC(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
-    u8* var_s1 = &frames[animFrameStart];
+    u8* iterEnt1 = &frames[animFrameStart];
     s16 var_a1 = 0;
 
     if (self->animFrameDuration == 0) {
-        if (*var_s1 != 0) {
-            if (*var_s1 == 0xFF) {
+        if (*iterEnt1 != 0) {
+            if (*iterEnt1 == 0xFF) {
                 return 0;
             }
-            self->animFrameDuration = *var_s1++ + (u8)self->ext.stub[0x3F];
-            self->animCurFrame = *var_s1++;
+            self->animFrameDuration = *iterEnt1++ + (u8)self->ext.stub[0x3F];
+            self->animCurFrame = *iterEnt1++;
             self->animFrameIdx++;
             var_a1 = 128;
         } else {
-            var_s1 = frames;
+            iterEnt1 = frames;
             self->animFrameIdx = 0;
             self->animFrameDuration = 0;
             self->ext.stub[0x3F] = (arg2 * Random()) >> 8;
-            self->animFrameDuration = *var_s1++ + (u8)self->ext.stub[0x3F];
-            self->animCurFrame = *var_s1;
+            self->animFrameDuration = *iterEnt1++ + (u8)self->ext.stub[0x3F];
+            self->animCurFrame = *iterEnt1;
             self->animFrameIdx++;
             return 0;
         }
     }
     self->animFrameDuration--;
-    self->animCurFrame = var_s1[-1];
+    self->animCurFrame = iterEnt1[-1];
     var_a1 |= 1;
     return var_a1;
 }
 
 // Absolute distance from g_CurrentEntity to the player in the X Axis
 s32 GetDistanceToPlayerX(void) {
-    s16 xDistance = g_CurrentEntity->posX.i.hi - PLAYER.posX.i.hi;
+    s16 yDistance = g_CurrentEntity->posX.i.hi - PLAYER.posX.i.hi;
 
-    if (xDistance < 0) {
-        xDistance = -xDistance;
+    if (yDistance < 0) {
+        yDistance = -yDistance;
     }
-    return xDistance;
+    return yDistance;
 }
 
 // Absolute distance from g_CurrentEntity to the player in the Y Axis
@@ -59,7 +55,7 @@ s32 GetDistanceToPlayerY(void) {
 
 #include "../get_side_to_player.h"
 
-void MoveEntity() {
+void MoveEntity(void) {
     g_CurrentEntity->posX.val += g_CurrentEntity->velocityX;
     g_CurrentEntity->posY.val += g_CurrentEntity->velocityY;
 }
@@ -70,7 +66,7 @@ void FallEntity(void) {
     }
 }
 
-s32 func_80194674(u16* sensors) {
+s32 func_801BC8E4(u16* sensors) {
     Collider col;
     Collider colBack;
     s16 x;
@@ -123,7 +119,7 @@ s32 func_80194674(u16* sensors) {
     return 0;
 }
 
-s32 func_801948EC(s16* posX) {
+s32 func_801BCB5C(s16* posX) {
     Collider collider;
     s16 temp2;
     s16 temp4;
@@ -191,39 +187,34 @@ Entity* AllocEntity(Entity* start, Entity* end) {
     return NULL;
 }
 
-s32 func_80194B34(u8 arg0, s16 arg1) { return D_80180BBC[arg0] * arg1; }
+s32 func_801BCDA4(u8 arg0, s16 arg1) { return D_80181A50[arg0] * arg1; }
 
-s16 func_80194B60(u8 arg0) { return D_80180BBC[arg0]; }
+s16 func_801BCDD0(u8 arg0) { return D_80181A50[arg0]; }
 
-void func_80194B7C(s32 arg0, s16 arg1) {
-    g_CurrentEntity->velocityX = func_80194B34(arg0 & 0xFF, arg1);
-    g_CurrentEntity->velocityY = func_80194B34((arg0 - 0x40) & 0xFF, arg1);
+void func_801BCDEC(s32 arg0, s16 arg1) {
+    g_CurrentEntity->velocityX = func_801BCDA4(arg0, arg1);
+    g_CurrentEntity->velocityY = func_801BCDA4(arg0 - 0x40, arg1);
 }
 
-u8 func_80194BE8(s16 x, s16 y) { return (ratan2(y, x) >> 4) + 0x40; }
+u8 func_801BCE58(s16 x, s16 y) { return (ratan2(y, x) >> 4) + 0x40; }
 
-u8 func_80194C20(Entity* arg0, Entity* arg1) {
-    u16 x;
-    u16 y;
-
-    x = arg1->posX.i.hi - arg0->posX.i.hi;
-    y = arg1->posY.i.hi - arg0->posY.i.hi;
-
-    return func_80194BE8(x, y);
+u8 func_801BCE90(Entity* a, Entity* b) {
+    s32 diffX = (u16)b->posX.i.hi - (u16)a->posX.i.hi;
+    s32 diffY = (u16)b->posY.i.hi - (u16)a->posY.i.hi;
+    return func_801BCE58(diffX, diffY);
 }
 
-u16 func_80194C68(s16 x, s16 y) {
-    x -= g_CurrentEntity->posX.i.hi;
-    y -= g_CurrentEntity->posY.i.hi;
-
-    return func_80194BE8(x, y);
+u8 func_801BCED8(s32 x, s32 y) {
+    s32 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
+    s32 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
+    return func_801BCE58(diffX, diffY);
 }
 
 #include "../adjust_value_within_threshold.h"
 
 #include "../unk_entity_func0.h"
 
-u16 func_80194D94(s16 arg0, s16 arg1) { return ratan2(arg1, arg0); }
+u16 func_801BD004(s16 x, s16 y) { return ratan2(y, x); }
 
 u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     s32 diffX = b->posX.i.hi - a->posX.i.hi;
@@ -231,13 +222,13 @@ u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     return ratan2(diffY, diffX);
 }
 
-u16 func_80194DFC(s32 x, s32 y) {
+u16 func_801BD06C(s32 x, s32 y) {
     s16 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
     s16 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
     return ratan2(diffY, diffX);
 }
 
-u16 GetNormalizedAngle(u16 arg0, u16 arg1, u16 arg2) {
+u16 GetNormalizedAngle(u16 arg0, s16 arg1, s16 arg2) {
     u16 var_v0 = arg1;
     u16 temp_a2 = arg2 - arg1;
     u16 var_v0_2;
@@ -262,59 +253,53 @@ u16 GetNormalizedAngle(u16 arg0, u16 arg1, u16 arg2) {
 }
 
 void SetStep(u8 step) {
-    Entity* entity = g_CurrentEntity;
-
-    entity->step = step;
-    entity->step_s = 0;
-    entity->animFrameIdx = 0;
-    entity->animFrameDuration = 0;
+    g_CurrentEntity->step = step;
+    g_CurrentEntity->step_s = 0;
+    g_CurrentEntity->animFrameIdx = 0;
+    g_CurrentEntity->animFrameDuration = 0;
 }
 
-void func_80194EC4(u8 arg0) {
-    Entity* entity = g_CurrentEntity;
-
-    entity->step_s = arg0;
-    entity->animFrameIdx = 0;
-    entity->animFrameDuration = 0;
+void SetSubStep(u8 step_s) {
+    g_CurrentEntity->step_s = step_s;
+    g_CurrentEntity->animFrameIdx = 0;
+    g_CurrentEntity->animFrameDuration = 0;
 }
 
-void EntityExplosionSpawn(u16 arg0, u16 arg1) {
-    Entity* entity;
-
-    if (arg1 != 0) {
-        func_8019A328(arg1);
+void EntityExplosionSpawn(u16 arg0, u16 sfxId) {
+    if (sfxId != 0) {
+        func_801C2598(sfxId);
     }
+
     if (arg0 == 0xFF) {
         DestroyEntity(g_CurrentEntity);
         return;
     }
-    entity = g_CurrentEntity;
-    entity->drawFlags = 0;
-    entity->entityId = 2;
-    entity->pfnUpdate = EntityExplosion;
-    entity->params = arg0;
-    entity->animCurFrame = 0;
+
+    g_CurrentEntity->entityId = E_EXPLOSION;
+    g_CurrentEntity->pfnUpdate = (PfnEntityUpdate)EntityExplosion;
+    g_CurrentEntity->params = arg0;
+    g_CurrentEntity->animCurFrame = 0;
+    g_CurrentEntity->drawFlags = 0;
     g_CurrentEntity->step = 0;
     g_CurrentEntity->step_s = 0;
 }
 
 #include "../init_entity.h"
 
-void func_80195070(Entity* entity) {
-    if (entity->step == 0) {
-        entity->step++;
+void EntityDummy(Entity* arg0) {
+    if (arg0->step == 0) {
+        arg0->step++;
     }
 }
 
-s32 func_80195098(u16* hitSensors, s16 sensorCount) {
+s32 func_801BD308(u16* hitSensors, s16 sensorCount) {
+    s32 velocityX = g_CurrentEntity->velocityX;
     Collider collider;
-    s16 i;
-    s32 velocityX;
     u16 temp_a1;
     s16 x;
     s16 y;
+    s16 i;
 
-    velocityX = g_CurrentEntity->velocityX;
     if (velocityX != 0) {
         x = g_CurrentEntity->posX.i.hi;
         y = g_CurrentEntity->posY.i.hi;
@@ -338,18 +323,19 @@ s32 func_80195098(u16* hitSensors, s16 sensorCount) {
     }
 }
 
-void func_801951C0(u16* hitSensors, s16 sensorCount) {
+void func_801BD430(u16* hitSensors, s16 sensorCount) {
+    s32 velocityX = g_CurrentEntity->velocityX;
     Collider collider;
-    s16 i;
-    s32 velocityX;
     s16 x;
     s16 y;
+    s16 i;
 
-    velocityX = g_CurrentEntity->velocityX;
     if (velocityX == 0)
         return;
+
     x = g_CurrentEntity->posX.i.hi;
     y = g_CurrentEntity->posY.i.hi;
+
     for (i = 0; i < sensorCount; i++) {
         if (velocityX < 0) {
             x = x + *hitSensors++;
@@ -375,7 +361,7 @@ void func_801951C0(u16* hitSensors, s16 sensorCount) {
 
 #include "../replace_breakable_with_item_drop.h"
 
-void func_80195714(void) {
+void func_801BD984(void) {
     s32 temp_v1;
     Entity* entity;
 
@@ -397,7 +383,7 @@ void func_80195714(void) {
     }
 }
 
-void func_80195798(u16 arg0) {
+void func_801BDA08(u16 arg0) {
     Collider collider;
 
     if (g_CurrentEntity->velocityX < 0) {
@@ -423,16 +409,14 @@ void func_80195798(u16 arg0) {
 
         if (collider.effects & EFFECT_QUICKSAND) {
             g_CurrentEntity->posY.val += FIX(0.125);
-            return;
+        } else {
+            g_CurrentEntity->posY.i.hi += collider.unk18;
         }
-
-        g_CurrentEntity->posY.i.hi += collider.unk18;
-        return;
-    }
-
-    if (!(collider.effects & EFFECT_NOTHROUGH)) {
-        MoveEntity();
-        func_80195714();
+    } else {
+        if (!(collider.effects & EFFECT_NOTHROUGH)) {
+            MoveEntity();
+            func_801BD984();
+        }
     }
 }
 
@@ -446,13 +430,13 @@ void CollectSubweapon(u16 subWeaponIdx) {
 
     g_api.PlaySfx(NA_SE_PL_IT_PICKUP);
     subWeapon = g_Status.subWeapon;
-    g_Status.subWeapon = D_80180EEC[subWeaponIdx];
+    g_Status.subWeapon = D_80181D80[subWeaponIdx];
 
     if (subWeapon == g_Status.subWeapon) {
         subWeapon = 1;
         g_CurrentEntity->unk6D[0] = 0x10;
     } else {
-        subWeapon = D_80180F1C[subWeapon];
+        subWeapon = D_80181DB0[subWeapon];
         g_CurrentEntity->unk6D[0] = 0x60;
     }
 
@@ -475,22 +459,10 @@ void CollectSubweapon(u16 subWeaponIdx) {
 
 #include "../collect_heart_vessel.h"
 
-void func_80195C0C(void) {
+void CollectLifeVessel(void) {
     g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
     g_api.func_800FE044(5, 0x8000);
     DestroyEntity(g_CurrentEntity);
 }
 
-void func_80195C5C(void) { DestroyEntity(g_CurrentEntity); }
-
-INCLUDE_ASM("st/cen/nonmatchings/12A3C", EntityPrizeDrop);
-
-#define MISSING_ANIMATE_ENTITY_PROTOTYPE
-#include "../entity_explosion.h"
-#undef MISSING_ANIMATE_ENTITY_PROTOTYPE
-
-#include "../blink_item.h"
-
-INCLUDE_ASM("st/cen/nonmatchings/12A3C", EntityEquipItemDrop);
-
-#include "../blit_char.h"
+void DestroyCurrentEntity(void) { DestroyEntity(g_CurrentEntity); }
