+++ src/st/nz0/create_entity.c
@@ -1,0 +1,3 @@
+#include <stage.h>
+
+#include "../libstage.h"
