+++ src/st/rwrp/BD58.c
@@ -2,7 +2,26 @@
 
 #include "../libstage.h"
 
-#include "../entity_is_near_player.h"
+bool EntityIsNearPlayer(Entity* self) {
+    s16 distanceX;
+    s16 diffX;
+    s16 distanceY;
+    s16 diffY;
+
+    diffX = PLAYER.posX.i.hi - self->posX.i.hi;
+    distanceX = abs(diffX);
+    if (distanceX > 16) {
+        return false;
+    }
+
+    diffY = PLAYER.posY.i.hi - self->posY.i.hi;
+    distanceY = abs(diffY);
+    if (distanceY > 32) {
+        return false;
+    }
+
+    return true;
+}
 
 INCLUDE_ASM("st/rwrp/nonmatchings/BD58", func_8018C948);
