+++ src/st/st0/34908.c
@@ -1,12 +1,8 @@
-#include "nz0.h"
-
-#include "../libstage.h"
-
-#include "../entity_red_door.h"
+#include "st0.h"
 
 #include "../entity.h"
 
-u8 func_801BCAD4(u8 frames[], Entity* self, u8 arg2) {
+u8 func_801B4AF0(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
     u8* var_s1 = &frames[animFrameStart];
     s16 var_a1 = 0;
@@ -38,7 +34,7 @@ u8 func_801BCAD4(u8 frames[], Entity* self, u8 arg2) {
 }
 
 // Absolute distance from g_CurrentEntity to the player in the X Axis
-s32 GetDistanceToPlayerX(void) {
+s16 GetDistanceToPlayerX(void) {
     s16 xDistance = g_CurrentEntity->posX.i.hi - PLAYER.posX.i.hi;
 
     if (xDistance < 0) {
@@ -57,9 +53,16 @@ s32 GetDistanceToPlayerY(void) {
     return yDistance;
 }
 
-#include "../get_side_to_player.h"
+s32 GetSideToPlayer(void) {
+    s16 side = g_CurrentEntity->posX.i.hi > PLAYER.posX.i.hi;
 
-void MoveEntity() {
+    if (g_CurrentEntity->posY.i.hi > PLAYER.posY.i.hi) {
+        side |= 2;
+    }
+    return side;
+}
+
+void MoveEntity(void) {
     g_CurrentEntity->posX.val += g_CurrentEntity->velocityX;
     g_CurrentEntity->posY.val += g_CurrentEntity->velocityY;
 }
@@ -70,7 +73,23 @@ void FallEntity(void) {
     }
 }
 
-s32 func_801BCCFC(u16* sensors) {
+u8 func_801B4D18(void) {
+    u8 unkState;
+    Entity* entity;
+
+    MoveEntity();
+    FallEntity();
+
+    entity = g_CurrentEntity;
+
+    if (unkState & 1) {
+        entity->velocityY = 0;
+    }
+
+    return unkState;
+}
+
+s32 func_801B4D5C(u16* sensors) {
     Collider col;
     Collider colBack;
     s16 x;
@@ -123,64 +142,12 @@ s32 func_801BCCFC(u16* sensors) {
     return 0;
 }
 
-s32 func_801BCF74(s16* posX) {
-    Collider collider;
-    s16 temp2;
-    s16 temp4;
-    s16 x, y;
+INCLUDE_ASM("st/st0/nonmatchings/34908", func_801B4FD4);
 
-    g_CurrentEntity->posX.val += g_CurrentEntity->velocityX;
-    temp2 = g_CurrentEntity->posY.i.hi + 3;
-    g_CurrentEntity->posY.i.hi = temp2;
-    x = g_CurrentEntity->posX.i.hi + *posX;
-    posX++;
-    y = temp2 + *posX;
-    g_api.CheckCollision(x, y, &collider, 0);
-    if (!(collider.effects & EFFECT_SOLID)) {
-        return 0;
-    }
-    posX++;
-
-    g_CurrentEntity->posY.i.hi = g_CurrentEntity->posY.i.hi + collider.unk18;
-    if (g_CurrentEntity->velocityX != 0) {
-        if (g_CurrentEntity->velocityX < 0) {
-            temp4 = x - *posX;
-            posX++;
-        } else {
-            temp4 = x + *posX;
-            posX++;
-        }
-        y = y + *posX - 7;
-        g_api.CheckCollision(temp4, y, &collider, 0);
-        if (collider.effects & EFFECT_SOLID) {
-            if ((collider.effects & (EFFECT_UNK_8000 | EFFECT_UNK_0002)) ==
-                EFFECT_UNK_0002) {
-                g_CurrentEntity->posX.val =
-                    g_CurrentEntity->posX.val - g_CurrentEntity->velocityX;
-                g_CurrentEntity->velocityX = 0;
-                return 0xFF;
-            }
-            return 0x61;
-        }
-        y += 15;
-        g_api.CheckCollision(temp4, y, &collider, 0);
-        if (collider.effects & EFFECT_SOLID) {
-            if (collider.effects & EFFECT_UNK_8000) {
-                return 0x61;
-            }
-            return 1;
-        }
-        g_CurrentEntity->posX.val -= g_CurrentEntity->velocityX;
-        g_CurrentEntity->velocityX = 0;
-
-        return 0x80;
-    }
-    return 1;
-}
+INCLUDE_ASM("st/st0/nonmatchings/34908", func_801B51E4);
 
 Entity* AllocEntity(Entity* start, Entity* end) {
     Entity* current = start;
-
     while (current < end) {
         if (current->entityId == E_NONE) {
             DestroyEntity(current);
@@ -192,40 +159,36 @@ Entity* AllocEntity(Entity* start, Entity* end) {
     return NULL;
 }
 
-s32 func_801BD1BC(u8 arg0, s16 arg1) { return D_80181978[arg0] * arg1; }
+s32 func_801B542C(u8 arg0, s16 arg1) { return D_80181990[arg0] * arg1; }
 
-s16 func_801BD1E8(u8 arg0) { return D_80181978[arg0]; } // Unique
+s16 func_801B5458(u8 arg0) { return D_80181990[arg0]; }
 
-void func_801BD204(s32 arg0, s16 arg1) {
-    g_CurrentEntity->velocityX = func_801BD1BC(arg0, arg1);
-    g_CurrentEntity->velocityY = func_801BD1BC(arg0 - 0x40, arg1);
+void func_801B5474(s32 arg0, s16 arg1) {
+    g_CurrentEntity->velocityX = func_801B542C(arg0, arg1);
+    g_CurrentEntity->velocityY = func_801B542C(arg0 - 0x40, arg1);
 }
 
-u8 func_801BD270(s16 x, s16 y) { return (ratan2(y, x) >> 4) + 0x40; }
-
-u8 func_8019AD64(ObjInit2* arg0, ObjInit2* arg1) {
-    u16 x, y;
-
-    x = arg1->zPriority - arg0->zPriority;
-    y = arg1->palette - arg0->palette;
-
-    return func_801BD270(x, y);
+u8 func_801B54E0(s16 arg0, s16 arg1) {
+    return (ratan2(arg1, arg0) >> 4) + 0x40;
 }
 
-u8 func_801BD2F0(s16 arg0, s16 arg1) {
-    s16 x, y;
+u8 func_801B5518(Entity* arg0, Entity* arg1) {
+    s16 a = arg1->posX.i.hi - arg0->posX.i.hi;
+    s16 b = arg1->posY.i.hi - arg0->posY.i.hi;
+    return func_801B54E0(a, b);
+}
 
-    x = arg0 - g_CurrentEntity->posX.i.hi;
-    y = arg1 - g_CurrentEntity->posY.i.hi;
-
-    return func_801BD270(x, y);
+u8 func_801B5560(s32 arg0, s32 arg1) {
+    s16 a = (arg0 - (u16)g_CurrentEntity->posX.i.hi);
+    s16 b = (arg1 - (u16)g_CurrentEntity->posY.i.hi);
+    return func_801B54E0(a, b);
 }
 
 #include "../adjust_value_within_threshold.h"
 
 #include "../unk_entity_func0.h"
 
-u16 func_801BD41C(s16 x, s16 y) { return ratan2(y, x); }
+u16 func_801B568C(s16 x, s16 y) { return ratan2(y, x); }
 
 u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     s32 diffX = b->posX.i.hi - a->posX.i.hi;
@@ -233,32 +196,30 @@ u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     return ratan2(diffY, diffX);
 }
 
-u16 func_801BD484(s32 x, s32 y) {
+u16 func_801B56F4(s32 x, s32 y) {
     s16 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
     s16 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
     return ratan2(diffY, diffX);
 }
 
-// TODO(sestren): Reconcile this GetNormalizedAngle with the one in 3E30C.c
 u16 GetNormalizedAngle(u16 arg0, s16 arg1, s16 arg2) {
-    u16 var_v0 = arg1;
     u16 temp_a2 = arg2 - arg1;
-    u16 var_v0_2;
+    u16 ret;
 
     if (temp_a2 & 0x800) {
-        var_v0_2 = (0x800 - temp_a2) & 0x7FF;
+        ret = temp_a2 & 0x7FF;
     } else {
-        var_v0_2 = temp_a2;
+        ret = temp_a2;
     }
 
-    if (var_v0_2 > arg0) {
+    if (ret > arg0) {
         if (temp_a2 & 0x800) {
-            var_v0 = arg1 - arg0;
+            ret = arg1 - arg0;
         } else {
-            var_v0 = arg1 + arg0;
+            ret = arg1 + arg0;
         }
 
-        return var_v0;
+        return ret;
     }
 
     return arg2;
@@ -277,53 +238,24 @@ void SetSubStep(u8 step_s) {
     g_CurrentEntity->animFrameDuration = 0;
 }
 
-void EntityExplosionSpawn(u16 arg0, u16 sfxId) {
-    if (sfxId != 0) {
-        func_801C29B0(sfxId);
-    }
+void func_801B57D0(u16 params) {
+    Entity* current;
 
-    if (arg0 == 0xFF) {
+    if (params == 0xFF) {
         DestroyEntity(g_CurrentEntity);
         return;
     }
-
+    current = g_CurrentEntity;
     g_CurrentEntity->entityId = E_EXPLOSION;
-    g_CurrentEntity->pfnUpdate = (PfnEntityUpdate)EntityExplosion;
-    g_CurrentEntity->params = arg0;
-    g_CurrentEntity->animCurFrame = 0;
+    g_CurrentEntity->pfnUpdate = EntityExplosion;
     g_CurrentEntity->drawFlags = 0;
+    current->params = params;
+    current->animCurFrame = 0;
     g_CurrentEntity->step = 0;
     g_CurrentEntity->step_s = 0;
 }
 
-void InitializeEntity(u16 arg0[]) {
-    u16 enemyId;
-    EnemyDef* enemyDef;
-
-    g_CurrentEntity->animSet = *arg0++;
-    g_CurrentEntity->animCurFrame = *arg0++;
-    g_CurrentEntity->unk5A = *arg0++;
-    g_CurrentEntity->palette = *arg0++;
-
-    enemyId = *arg0++;
-    g_CurrentEntity->enemyId = enemyId;
-    enemyDef = &g_api.enemyDefs[enemyId];
-    g_CurrentEntity->hitPoints = enemyDef->hitPoints;
-    g_CurrentEntity->attack = enemyDef->attack;
-    g_CurrentEntity->attackElement = enemyDef->attackElement;
-    g_CurrentEntity->hitboxState = enemyDef->hitboxState;
-    g_CurrentEntity->hitboxWidth = enemyDef->hitboxWidth;
-    g_CurrentEntity->hitboxHeight = enemyDef->hitboxHeight;
-    g_CurrentEntity->flags = enemyDef->flags;
-    g_CurrentEntity->hitboxOffX = 0;
-    g_CurrentEntity->hitboxOffY = 0;
-    g_CurrentEntity->step_s = 0;
-    g_CurrentEntity->step++;
-    if (g_CurrentEntity->zPriority == 0) {
-        g_CurrentEntity->zPriority =
-            g_unkGraphicsStruct.g_zEntityCenter.unk - 0xC;
-    }
-}
+#include "../init_entity.h"
 
 void EntityDummy(Entity* arg0) {
     if (arg0->step == 0) {
@@ -331,7 +263,7 @@ void EntityDummy(Entity* arg0) {
     }
 }
 
-s32 func_801BD720(u16* hitSensors, s16 sensorCount) {
+s32 func_801B5970(u16* hitSensors, s16 sensorCount) {
     Collider collider;
     s16 i;
     s32 velocityX;
@@ -354,8 +286,8 @@ s32 func_801BD720(u16* hitSensors, s16 sensorCount) {
 
             y += *hitSensors++;
             g_api.CheckCollision(x, y, &collider, 0);
-            if (collider.effects & EFFECT_UNK_0002 &&
-                ((!(collider.effects & EFFECT_UNK_8000)) || (i != 0))) {
+            if (collider.effects & 2 &&
+                ((!(collider.effects & 0x8000)) || (i != 0))) {
                 return 2;
             }
         }
@@ -363,31 +295,28 @@ s32 func_801BD720(u16* hitSensors, s16 sensorCount) {
     }
 }
 
-void CheckFieldCollision(u16* hitSensors, s16 sensorCount) {
-    s32 velocityX = g_CurrentEntity->velocityX;
+void func_801B5A98(u16* hitSensors, s16 sensorCount) {
     Collider collider;
-    s16 x, y;
     s16 i;
+    s32 velocityX;
+    s16 x;
+    s16 y;
 
-    if (velocityX == 0) {
+    velocityX = g_CurrentEntity->velocityX;
+    if (velocityX == 0)
         return;
-    }
-
     x = g_CurrentEntity->posX.i.hi;
     y = g_CurrentEntity->posY.i.hi;
-
     for (i = 0; i < sensorCount; i++) {
         if (velocityX < 0) {
-            x += *hitSensors++;
+            x = x + *hitSensors++;
         } else {
-            x -= *hitSensors++;
+            x = x - *hitSensors++;
         }
+
         y += *hitSensors++;
-
         g_api.CheckCollision(x, y, &collider, 0);
-
-        if ((collider.effects & EFFECT_UNK_0002) &&
-            (!(collider.effects & EFFECT_UNK_8000) || (i != 0))) {
+        if (collider.effects & 2 && (!(collider.effects & 0x8000) || i != 0)) {
             if (velocityX < 0) {
                 g_CurrentEntity->posX.i.hi += collider.unk1C;
             } else {
@@ -398,6 +327,129 @@ void CheckFieldCollision(u16* hitSensors, s16 sensorCount) {
     }
 }
 
-#include "../get_player_collision_with.h"
+INCLUDE_ASM("st/st0/nonmatchings/34908", func_801B5BF0);
 
 #include "../replace_breakable_with_item_drop.h"
+
+void func_801B5EC8(void) {
+    s32 temp_v1;
+    Entity* entity;
+
+    entity = g_CurrentEntity;
+    if (entity->velocityY >= 0) {
+        temp_v1 =
+            entity->ext.generic.unk88.S16.unk0 + entity->ext.generic.unk84.unk;
+        entity->ext.generic.unk84.unk = temp_v1;
+        entity->velocityX = temp_v1;
+        if (temp_v1 == 0x10000 || temp_v1 == -0x10000) {
+            entity->ext.generic.unk88.S16.unk0 =
+                -entity->ext.generic.unk88.S16.unk0;
+        }
+        entity = g_CurrentEntity;
+    }
+
+    if (entity->velocityY < FIX(0.25)) {
+        entity->velocityY += FIX(0.125);
+    }
+}
+
+void func_801B5F4C(u16 arg0) {
+    Collider res;
+
+    if (g_CurrentEntity->velocityX < 0) {
+        g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
+                             g_CurrentEntity->posY.i.hi - 7, &res, 0);
+        if (res.effects & 5) {
+            g_CurrentEntity->velocityY = 0;
+        }
+    }
+
+    g_api.CheckCollision(
+        g_CurrentEntity->posX.i.hi, g_CurrentEntity->posY.i.hi + 7, &res, 0);
+
+    if (arg0) {
+        if (!(res.effects & 5)) {
+            MoveEntity();
+            FallEntity();
+            return;
+        }
+
+        g_CurrentEntity->velocityX = 0;
+        g_CurrentEntity->velocityY = 0;
+
+        if (res.effects & 4) {
+            g_CurrentEntity->posY.val += FIX(0.125);
+            return;
+        }
+
+        g_CurrentEntity->posY.i.hi += res.unk18;
+        return;
+    }
+
+    if (!(res.effects & 5)) {
+        MoveEntity();
+        func_801B5EC8();
+    }
+}
+
+// Can't be deduped with the other instances due to the extra memcpy
+void CollectHeart(u16 index) {
+    s8 hearts[10];
+
+    __builtin_memcpy(&hearts[0], &c_HeartPrizes[0], 2);
+    __builtin_memcpy(&hearts[8], &c_HeartPrizes[4], 2);
+    g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
+    g_Status.hearts = hearts[index] + g_Status.hearts;
+    if (g_Status.heartsMax < g_Status.hearts) {
+        g_Status.hearts = g_Status.heartsMax;
+    }
+    DestroyEntity(g_CurrentEntity);
+}
+
+#include "../collect_gold.h"
+
+void func_801B61D4(void) {}
+
+void func_801B61DC(void) { DestroyEntity(g_CurrentEntity); }
+
+void CollectSubweapon(u16 subWeaponIdx) {
+    Entity* player = &PLAYER;
+    u16 subWeapon;
+
+    g_api.PlaySfx(NA_SE_PL_IT_PICKUP);
+    subWeapon = g_Status.subWeapon;
+    g_Status.subWeapon = D_80181CAC[subWeaponIdx];
+
+    if (subWeapon == g_Status.subWeapon) {
+        subWeapon = 1;
+        g_CurrentEntity->unk6D[0] = 0x10;
+    } else {
+        subWeapon = D_80181CDC[subWeapon];
+        g_CurrentEntity->unk6D[0] = 0x60;
+    }
+
+    if (subWeapon != 0) {
+        g_CurrentEntity->params = subWeapon;
+        g_CurrentEntity->posY.i.hi = player->posY.i.hi + 12;
+        g_CurrentEntity->step = 7;
+        g_CurrentEntity->step_s = 0;
+        g_CurrentEntity->velocityY = FIX(-2.5);
+        g_CurrentEntity->animCurFrame = 0;
+        g_CurrentEntity->ext.generic.unk88.S16.unk2 = 5;
+        if (player->facingLeft != 1) {
+            g_CurrentEntity->velocityX = FIX(-2);
+            return;
+        }
+        g_CurrentEntity->velocityX = FIX(2);
+        return;
+    }
+    DestroyEntity(g_CurrentEntity);
+}
+
+void CollectDummy(void) { DestroyEntity(g_CurrentEntity); }
+
+Entity* func_801B633C(void) {
+    g_CurrentEntity->step = 3;
+    g_CurrentEntity->params = 4;
+    return g_CurrentEntity;
+}
