+++ src/st/st0/create_entity.c
@@ -1,0 +1,1 @@
+#include "../libstage.h"
