+++ src/st/wrp_psp/e_red_door.c
@@ -1,6 +1,25 @@
 #include "../wrp/wrp.h"
 
-#include "../entity_is_near_player.h"
+bool EntityIsNearPlayer(Entity* self) {
+    s16 distanceX;
+    s16 diffX;
+    s16 distanceY;
+    s16 diffY;
+
+    diffX = PLAYER.posX.i.hi - self->posX.i.hi;
+    distanceX = abs(diffX);
+    if (distanceX > 16) {
+        return false;
+    }
+
+    diffY = PLAYER.posY.i.hi - self->posY.i.hi;
+    distanceY = abs(diffY);
+    if (distanceY > 32) {
+        return false;
+    }
+
+    return true;
+}
 
 extern u8 g_eRedDoorUV[3][8];
 extern u16 g_eRedDoorTiles[2][8];
